Split Preconditions into Core (Check) and Java.Search (UsesType/UsesMethod)
Move language-agnostic Check class and Preconditions.Check() to
OpenRewrite.Core namespace. Keep UsesType/UsesMethod in
OpenRewrite.Java.Search since they depend on RPC and Java-specific
search. Callers use static imports to access both without ambiguity.

Also fix xUnit1031 warnings in PreconditionsCheckTest by making
tests async.
